+++ elixir/README.md
@@ -262,7 +262,7 @@ Steps:
 
    ```bash
    cd elixir/
-   BILLING_ENABLED=true STRIPE_SECRET_KEY="...copy from stripe dashboard..." STRIPE_WEBHOOK_SIGNING_SECRET="...copy from stripe cli tool.." mix phx.server
+   STRIPE_SECRET_KEY="...copy from stripe dashboard..." STRIPE_WEBHOOK_SIGNING_SECRET="...copy from stripe cli tool.." mix phx.server
    ```
 
 When updating the billing plan in stripe, use the
