+++ elixir/apps/domain/lib/domain/billing.ex
@@ -17,98 +17,19 @@ defmodule Domain.Billing do
       Jobs.CheckAccountLimits
     ]
 
-    if enabled?() do
-      Supervisor.init(children, strategy: :one_for_one)
-    else
-      :ignore
-    end
+    Supervisor.init(children, strategy: :one_for_one)
   end
 
   # Configuration helpers
-
-  def enabled? do
-    fetch_config!(:enabled)
-  end
 
   def fetch_webhook_signing_secret! do
     fetch_config!(:webhook_signing_secret)
   end
 
-  # Limits and Features
-
-  def users_limit_exceeded?(%Accounts.Account{} = account, users_count) do
-    not is_nil(account.limits.users_count) and
-      users_count > account.limits.users_count
-  end
-
-  def seats_limit_exceeded?(%Accounts.Account{} = account, active_users_count) do
-    not is_nil(account.limits.monthly_active_users_count) and
-      active_users_count > account.limits.monthly_active_users_count
-  end
-
-  def can_create_users?(%Accounts.Account{} = account) do
-    users_count = Actors.count_users_for_account(account)
-    active_users_count = Clients.count_1m_active_users_for_account(account)
-
-    cond do
-      not Accounts.account_active?(account) ->
-        false
-
-      not is_nil(account.limits.monthly_active_users_count) ->
-        active_users_count < account.limits.monthly_active_users_count
-
-      not is_nil(account.limits.users_count) ->
-        users_count < account.limits.users_count
-
-      true ->
-        true
-    end
-  end
-
-  def service_accounts_limit_exceeded?(%Accounts.Account{} = account, service_accounts_count) do
-    not is_nil(account.limits.service_accounts_count) and
-      service_accounts_count > account.limits.service_accounts_count
-  end
-
-  def can_create_service_accounts?(%Accounts.Account{} = account) do
-    service_accounts_count = Actors.count_service_accounts_for_account(account)
-
-    Accounts.account_active?(account) and
-      (is_nil(account.limits.service_accounts_count) or
-         service_accounts_count < account.limits.service_accounts_count)
-  end
-
-  def gateway_groups_limit_exceeded?(%Accounts.Account{} = account, gateway_groups_count) do
-    not is_nil(account.limits.gateway_groups_count) and
-      gateway_groups_count > account.limits.gateway_groups_count
-  end
-
-  def can_create_gateway_groups?(%Accounts.Account{} = account) do
-    gateway_groups_count = Gateways.count_groups_for_account(account)
-
-    Accounts.account_active?(account) and
-      (is_nil(account.limits.gateway_groups_count) or
-         gateway_groups_count < account.limits.gateway_groups_count)
-  end
-
-  def admins_limit_exceeded?(%Accounts.Account{} = account, account_admins_count) do
-    not is_nil(account.limits.account_admin_users_count) and
-      account_admins_count > account.limits.account_admin_users_count
-  end
-
-  def can_create_admin_users?(%Accounts.Account{} = account) do
-    account_admins_count = Actors.count_account_admin_users_for_account(account)
-
-    Accounts.account_active?(account) and
-      (is_nil(account.limits.account_admin_users_count) or
-         account_admins_count < account.limits.account_admin_users_count)
-  end
-
   # API wrappers
 
-  def create_customer(%Accounts.Account{} = account) do
+  def create_customer(%Accounts.Account{} = account, attrs) do
     secret_key = fetch_config!(:secret_key)
-    email = get_customer_email(account)
 
     with {:ok, %{"id" => customer_id, "email" => customer_email}} <-
            APIClient.create_customer(secret_key, account.legal_name, email, %{
@@ -144,14 +65,13 @@ defmodule Domain.Billing do
   defp get_customer_email(%{metadata: %{stripe: %{billing_email: email}}}), do: email
   defp get_customer_email(_account), do: nil
 
-  def update_customer(%Accounts.Account{} = account) do
+  def update_customer(%Accounts.Account{stripe_customer_id: stripe_customer_id} = account) do
     secret_key = fetch_config!(:secret_key)
-    customer_id = account.metadata.stripe.customer_id
 
     with {:ok, _customer} <-
            APIClient.update_customer(
              secret_key,
-             customer_id,
+             stripe_customer_id,
              account.legal_name,
              %{
                account_id: account.id,
@@ -198,7 +118,7 @@ defmodule Domain.Billing do
 
       {:ok, params} ->
         :ok =
-          Logger.info("Stripe customer does not have account_id in metadata",
+          Logger.warning("Stripe customer does not have account_id in metadata",
             customer_id: customer_id,
             metadata: inspect(params["metadata"])
           )
@@ -250,6 +170,33 @@ defmodule Domain.Billing do
     end
   end
 
+  def fetch_billing_info(%Accounts.Account{} = account) do
+    secret_key = fetch_config!(:secret_key)
+    customer_id = account.stripe_customer_id
+
+    with {:ok, %{"subscriptions" => subscriptions}} <-
+           APIClient.fetch_billing_info(secret_key, customer_id) do
+      {:ok, subscriptions}
+    else
+      {:error, {status, body}} ->
+        :ok =
+          Logger.error("Cannot fetch Stripe billing info",
+            status: status,
+            body: inspect(body)
+          )
+
+        {:error, :retry_later}
+
+      {:error, reason} ->
+        :ok =
+          Logger.error("Cannot fetch Stripe billing info",
+            reason: inspect(reason)
+          )
+
+        {:error, :retry_later}
+    end
+  end
+
   def fetch_product(product_id) do
     secret_key = fetch_config!(:secret_key)
 
@@ -277,19 +224,8 @@ defmodule Domain.Billing do
 
   # Account management, sync and provisioning
 
-  def account_provisioned?(%Accounts.Account{metadata: %{stripe: %{customer_id: customer_id}}})
-      when not is_nil(customer_id) do
-    enabled?()
-  end
-
-  def account_provisioned?(%Accounts.Account{}) do
-    false
-  end
-
   def provision_account(%Accounts.Account{} = account) do
-    with true <- enabled?(),
-         true <- not account_provisioned?(account),
-         {:ok, account} <- create_customer(account),
+    with {:ok, account} <- create_customer(account),
          {:ok, account} <- create_subscription(account) do
       {:ok, account}
     else
@@ -306,12 +242,6 @@ defmodule Domain.Billing do
     slug_changed? = Ecto.Changeset.changed?(changeset, :slug)
 
     cond do
-      not account_provisioned?(account) ->
-        :ok
-
-      not enabled?() ->
-        :ok
-
       not name_changed? and not slug_changed? ->
         :ok
