+++ elixir/apps/web/lib/web/live/settings/billing.ex
@@ -4,28 +4,27 @@ defmodule Web.Settings.Billing do
   require Logger
 
   def mount(_params, _session, socket) do
-    if Billing.account_provisioned?(socket.assigns.account) do
-      admins_count = Actors.count_account_admin_users_for_account(socket.assigns.account)
-      service_accounts_count = Actors.count_service_accounts_for_account(socket.assigns.account)
-      users_count = Actors.count_users_for_account(socket.assigns.account)
-      active_users_count = Clients.count_1m_active_users_for_account(socket.assigns.account)
-      gateway_groups_count = Gateways.count_groups_for_account(socket.assigns.account)
+    admins_count = Actors.count_account_admin_users_for_account(socket.assigns.account)
+    service_accounts_count = Actors.count_service_accounts_for_account(socket.assigns.account)
+    users_count = Actors.count_users_for_account(socket.assigns.account)
+    active_users_count = Clients.count_1m_active_users_for_account(socket.assigns.account)
+    gateway_groups_count = Gateways.count_groups_for_account(socket.assigns.account)
 
-      socket =
-        assign(socket,
-          page_title: "Billing",
-          error: nil,
-          admins_count: admins_count,
-          users_count: users_count,
-          active_users_count: active_users_count,
-          service_accounts_count: service_accounts_count,
-          gateway_groups_count: gateway_groups_count
-        )
+    billing_info = socket.assigns.account.metadata.stripe
 
-      {:ok, socket}
-    else
-      raise Web.LiveErrors.NotFoundError
-    end
+    socket =
+      assign(socket,
+        page_title: "Billing",
+        error: nil,
+        admins_count: admins_count,
+        users_count: users_count,
+        active_users_count: active_users_count,
+        service_accounts_count: service_accounts_count,
+        gateway_groups_count: gateway_groups_count,
+        billing_info: billing_info
+      )
+
+    {:ok, socket}
   end
 
   def render(assigns) do
@@ -34,7 +33,16 @@ defmodule Web.Settings.Billing do
       <.breadcrumb path={~p"/#{@account}/settings/billing"}>Billing</.breadcrumb>
     </.breadcrumbs>
 
-    <.section>
+    <.section :if={is_nil(@billing_info)}>
+      <:title>
+        Plan
+      </:title>
+      <:content>
+        Upgrade your plan to access more features.
+      </:content>
+    </.section>
+
+    <.section :if={not is_nil(@billing_info)}>
       <:title>
         Billing Information
       </:title>
@@ -65,7 +73,7 @@ defmodule Web.Settings.Billing do
           <.vertical_table_row>
             <:label>Current Plan</:label>
             <:value>
-              {@account.metadata.stripe.product_name}
+              {Accounts.current_plan(@account)}
               <span class="ml-1">
                 <.link
                   class={link_style()}
@@ -87,14 +95,14 @@ defmodule Web.Settings.Billing do
           <.vertical_table_row>
             <:label>Billing Email</:label>
             <:value>
-              {@account.metadata.stripe.billing_email}
+              {@billing_info.billing_email}
             </:value>
           </.vertical_table_row>
 
           <.vertical_table_row>
             <:label>Billing Name</:label>
             <:value>
-              {@account.legal_name}
+              {@billing_info.billing_name}
             </:value>
           </.vertical_table_row>
         </.vertical_table>
@@ -197,7 +205,7 @@ defmodule Web.Settings.Billing do
       </:title>
       <:content>
         <div class="ml-4 mb-4 text-neutral-600">
-          <span :if={@account.metadata.stripe.support_type == "email"}>
+          <span :if={not is_nil(@billing_info) and @billing_info.support_type == "email"}>
             Please send
             <.link
               class={link_style()}
@@ -215,7 +223,7 @@ defmodule Web.Settings.Billing do
             and we will get back to you as soon as possible.
           </span>
 
-          <span :if={@account.metadata.stripe.support_type == "email_and_slack"}>
+          <span :if={not is_nil(@billing_info) and @billing_info.support_type == "email_and_slack"}>
             Please send us a message in the shared Slack channel or <.link
               class={link_style()}
               target="_blank"
@@ -229,7 +237,9 @@ defmodule Web.Settings.Billing do
             >an email</.link>.
           </span>
 
-          <span :if={@account.metadata.stripe.support_type not in ["email", "email_and_slack"]}>
+          <span :if={
+            is_nil(@billing_info) or @billing_info.support_type not in ["email", "email_and_slack"]
+          }>
             Ask questions, get help from other Firezone users on
             <.link class={link_style()} href="https://discourse.firez.one/">
               Discourse
