+++ elixir/config/dev.exs
@@ -14,7 +14,6 @@ config :domain, Domain.Repo,
 config :domain, outbound_email_adapter_configured?: true
 
 config :domain, Domain.Billing,
-  enabled: System.get_env("BILLING_ENABLED", "false") == "true",
   secret_key: System.get_env("STRIPE_SECRET_KEY", "sk_dev_1111"),
   webhook_signing_secret: System.get_env("STRIPE_WEBHOOK_SIGNING_SECRET", "whsec_dev_1111")
